[Work in progress] A* implemented

+++ Real autochess/Assets/_Game/Scripts/Configs/AreaSettings.cs
@@ -8,8 +8,8 @@ namespace Scripts.Configs
 	{
 		[SerializeField] private int _width;
 		[SerializeField] private int _height;
-		[SerializeField] private int _straitCost = 10;
-		[SerializeField] private int _diagonalCost = 10;
+		[SerializeField] private float _straitCost = 10f;
+		[SerializeField] private float _diagonalCost = 10f;
 		[SerializeField] private Vector2 _nodeHalfSize;
 		[SerializeField] private Vector2 _spawnInxRange;
 		[SerializeField] private Color[] _tileColors;
@@ -17,8 +17,8 @@ namespace Scripts.Configs
 
 		public int Width => _width;
 		public int Height => _height;
-		public int StraitCost => _straitCost;
-		public int DiagonalCost => _diagonalCost;
+		public float StraitCost => _straitCost;
+		public float DiagonalCost => _diagonalCost;
 		public Color[] TileColors => _tileColors;
 		public Vector2 SpawnInxRange => _spawnInxRange;
 		public Vector2 NodeHalfSize => _nodeHalfSize;
